% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bi}
\alias{bi}
\title{Age-specific severity counts from Bi et al. 2020}
\format{
A data frame with 8 rows and 7 variables:
\describe{
  \item{age_group}{age group, in 10-year intervals}
  \item{mild}{number of mild cases}
  \item{moderate}{number of moderate cases}
  \item{severe}{number of severe cases}
  \item{fever_no}{number of cases with no fever}
  \item{fever_yes}{number of cases with fever}
  \item{total}{total number of cases}
}
}
\source{
Bi, Q., Wu, Y., ..., and Feng, .T. (2020) Epidemiology and
  Transmission of COVID-19 in Shenzhen China: Analysis of 391 cases and 1,286
  of their close contacts. medRxiv preprint.
  \url{https://doi.org/10.1101/2020.03.03.20028423}
}
\usage{
bi
}
\description{
Based on data from Shenzhen CDC, China.
}
\keyword{datasets}
