% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{build_priors}
\alias{build_priors}
\title{Helper function for manipulation of prior-related list-structures}
\usage{
build_priors(..., .postfix = c("_a", "_b"), .prefix = "")
}
\description{
Assumes each argument is a two-element vector. Captures the name of each
argument. Creates two list-elements for each argument, using the 
naming convention given by '.postfix'. Prefixes these list-element keys
with the string '.prefix'. Returns the new list.
}
