% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_age_severity.R
\name{compare_age_severity}
\alias{compare_age_severity}
\title{Compile age-specific estimates of Pr(hospitalizion|infection) from four
research groups}
\usage{
compare_age_severity(jhu_model = c("Update", "JHU Original"))
}
\arguments{
\item{jhu_model}{Which JHU model to use to derive posterior probabilities
("Update" or "JHU Original")}
}
\value{
A data.frame with age-specific estimates of Pr(hospitalization|infection)
(mean and 95% CI) based on methods by four different research groups (JHU,
Neher lab, Pasteur Institute, and LSHTM)
}
\description{
Compile age-specific estimates of Pr(hospitalizion|infection) based on
methods from four different research groups (JHU, Neher, Pasteur, and LSHTM).
}
\examples{
compare_age_severity()

}
\author{
Patrick Barks <patrick.barks@epicentre.msf.org>
}
