% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_IFR.R
\name{compare_IFR}
\alias{compare_IFR}
\title{Compile age-adjusted IFR estimates from different countries based on ensemble
estimates by O'Driscoll et al. 2020.}
\usage{
compare_IFR(countries = c("AFG", "JPN", "ITA", "FRA"))
}
\value{
A data.frame with IFR estimates (mean and 95% CI) for different countries
}
\description{
Compile age-adjusted IFR estimates from different countries based on ensemble
estimates by O'Driscoll et al. 2020.
}
\examples{
compare_IFR()

}
\author{
Flavio Finger <flavio.finger@epicentre.msf.org>
}
