% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{covidestim_add.priors}
\alias{covidestim_add.priors}
\title{When adding priors, we want to be sure that a new 'modelconfig' object is
created, in order to check these priors}
\usage{
\method{covidestim_add}{priors}(rightside, leftside)
}
\description{
When adding priors, we want to be sure that a new 'modelconfig' object is
created, in order to check these priors
}
