% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{davies}
\alias{davies}
\title{Age-specific clinical fraction estimates from Davies et al. 2020}
\format{
A data frame with 96 rows and 5 variables:
\describe{
  \item{variable}{variable, "p_clin_inf"}
  \item{age_group}{age group, in 10-year intervals}
  \item{stat}{statistic, "mean", "median", "low_50", "upp_50", "low_95", or "upp_95"}
  \item{probability}{probability clinical given infection}
  \item{quantile}{quantile corresponding to the relevant statistic}
}
}
\source{
Davies, N.G., Klepac, P., Liu, Y., Prem, K., Jit, M., CMMID COVID-19
  Working Group, and Eggo, R.M. (2020) Age-dependent effects in the
  transmission and control of COVID-19 epidemics. medRxiv preprint.
  \url{https://doi.org/10.1101/2020.03.24.20043018}
}
\usage{
davies
}
\description{
Estimates from Extended Data Table 1 of Davies et al. 2020
}
\keyword{datasets}
