% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconfig.R
\name{defaultConfig}
\alias{defaultConfig}
\title{High level description of the function}
\usage{
defaultConfig(...)
}
\arguments{
\item{param1}{Description}

\item{param2}{Description}

\item{param3}{Description}
}
\value{
The return value
}
\description{
More extended description of the function
}
\examples{
print(mtcars)
}
