% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{example_ct_data}
\alias{example_ct_data}
\title{Example CT case/death data}
\usage{
example_ct_data(type = "cases")
}
\arguments{
\item{type}{A string: \code{"cases"} or \code{"deaths"}}
}
\value{
A tibble with two variables: \code{date}, a vector of \code{Date}
objects, and \code{observation}, a vector of doubles.
}
\description{
Returns a \code{\link[tibble]{tibble}} of Connecticut case, death, vaccine RR,
booster and hospitalizations data from December 2, 2021 to June 30, 2022.
Only one type of data will be returned at a time.
}
\examples{

cfg <- covidestim(nweeks = 31, region = 'Connecticut',
   pop = get_pop("Connecticut")) +
  input_cases(example_ct_data('cases')) +
  input_deaths(example_ct_data('deaths')) 

\dontrun{
  result <- run(cfg)
}

}
