% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{example_nyc_data}
\alias{example_nyc_data}
\title{Example NYC case/death data}
\usage{
example_nyc_data(type = "cases")
}
\arguments{
\item{type}{A string: \code{"cases"} or \code{"deaths"}}
}
\value{
A tibble with two variables: \code{date}, a vector of \code{Date}
objects, and \code{observation}, a vector of doubles.
}
\description{
Returns a \code{\link[tibble]{tibble}} of NYC case and death
data from February 29, 2020 to June 27, 2020. Only one type of data will be
returned at a time.
}
\examples{

cfg <- covidestim(ndays = 120, region = 'New York') +
  input_cases(example_nyc_data('cases')) +
  input_deaths(example_nyc_data('deaths'))

\dontrun{
  result <- run(cfg)
}

}
