% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_pop.R
\name{get_age_pop}
\alias{get_age_pop}
\title{Extract age-specific population estimates for a given country from the UN
World Population Prospects 2019}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
get_age_pop(iso, format = c("wide", "long"))
}
\arguments{
\item{iso}{ISO3 code for country of interest}

\item{format}{Return format ("wide" or "long"). Format "wide" yields a
single-row data.frame with age classes as column names, whereas "long"
yields a 2-column data.frame with the first defining age classes and the
second giving corresponding population sizes. Defaults to "wide".}
}
\value{
A data.frame. If \code{format = "wide"} returns a single-row data.frame with
age classes as column names, whereas if \code{format = "long"} returns a
2-column data.frame with the first defining age classes and the second giving
corresponding population sizes.
}
\description{
Extract population estimates in 10-year age intervals for a given country
from the UN World Population Prospects 2019 (WPP2019).
}
\examples{
get_age_pop(iso = "FRA")
get_age_pop(iso = "FRA", format = "long")

}
