% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_est_brazeau}
\alias{get_est_brazeau}
\title{Get estimates from Brazeau}
\usage{
get_est_brazeau(type = "p_dead_inf", stat = NULL)
}
\arguments{
\item{type}{Optional type to subset to (either "p_dead_inf", "p_dead_inf_serorev").
Defaults to "p_dead_inf".}

\item{stat}{Optional statistic to subset to (either "mean", "low_95", or
"up_95"). Defaults to `NULL`, in which case all cagegories are returned.}
}
\description{
Get estimates from Brazeau
}
