% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_est_davies}
\alias{get_est_davies}
\title{Get estimates from Davies}
\usage{
get_est_davies(stat = NULL)
}
\arguments{
\item{stat}{Optional statistic to subset to (either "mean", "median",
"low_50", "up_50", "low_95", or "up_95"). Defaults to `NULL`, in which case
all statistics are returned.}
}
\description{
Get estimates from Davies
}
