% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_est_levin}
\alias{get_est_levin}
\title{Get estimates from Levin}
\usage{
get_est_levin(stat = NULL)
}
\arguments{
\item{stat}{Optional statistic to subset to (either "mean", "low_95", or
"up_95"). Defaults to `NULL`, in which case all cagegories are returned.}
}
\description{
Get estimates from Levin
}
