% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_ifr_prior}
\alias{get_ifr_prior}
\title{Estimated IFR priors for US counties and states}
\usage{
get_ifr_prior(region)
}
\arguments{
\item{region}{A string with the state name, or the FIPS code}
}
\value{
A two-element numeric vector, intended to be used with
  \code{\link{priors_transitions}} as the argument to \code{p_die_if_inf}.
}
\description{
Returns parameters of a gamma prior on the IFR, which is represented
internally as the probability of dying as if infectied. Look in `data-raw/`
for details on how this is generated.
}
\examples{
priors_transitions(p_die_if_inf = get_ifr_prior('Connecticut'))

}
