% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifr.R
\name{get_ifr_raw}
\alias{get_ifr_raw}
\title{Estimated IFR for US counties and states}
\usage{
\dontrun{
  get_ifr_raw('Connecticut')
  get_ifr_raw('09009')
}
}
\arguments{
\item{region}{A string with the state name, or the FIPS code}
}
\value{
A data.frame with two variables: \link{date, value}, where value is
the odds ratio on VALUE HERE for that particular day. If \code{region}
is specified incorrectly, an error will be thrown.
}
\description{
Look in \verb{data-raw/} for details on how this is generated.
}
