% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{get_imm_init}
\alias{get_imm_init}
\title{Immunity and cumulative infections (fraction) estimates for US states and counties}
\usage{
get_imm_init(region)
}
\arguments{
\item{region}{A string with the state name, or the FIPS code}
}
\value{
State/county population as a numeric, or an error
}
\description{
Returns December 1, 2021 covidestim estimates of state or county effective protection
and fraction of the population ever infected.
}
\examples{
get_imm_init('Connecticut')
get_imm_init('09009')

}
