% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_LHSTM.R
\name{get_p_LSHTM}
\alias{get_p_LSHTM}
\title{Get age specific probabilities estimated by LSHTM (Davies et al. 2020 and Van
Zandvoort et al. 2020)}
\usage{
get_p_LSHTM(
  age_distr,
  p_type = c("p_hosp_inf", "p_icu_hosp", "p_dead_hosp", "p_dead_inf"),
  p_stat = c("mean", "median", "low_95", "up_95", "low_50", "up_50")
)
}
\arguments{
\item{age_distr}{age distribution data-frame}

\item{p_type}{type of probablity to extract: p_hosp_inf = P(Hosp|Infection), etc}

\item{p_stat}{quantile of the original estimates used to compute probability}
}
\description{
Get age specific probabilities estimated by LSHTM (Davies et al. 2020 and Van
Zandvoort et al. 2020)
}
\examples{
# overall propability of hospitalization given infection for France
age_dist_fr <- get_age_pop(iso = "FRA", format = "long")
get_p_LSHTM(age_dist_fr, p_type = "p_hosp_inf", p_stat = "median")

}
\author{
Anton Camacho
}
