% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_Salje2020.R
\name{get_p_Salje2020}
\alias{get_p_Salje2020}
\title{Get age-sex specific probabilities estimated by Salje et al. 2020}
\usage{
get_p_Salje2020(
  age_distr,
  p_type = c("p_hosp_inf", "p_icu_hosp", "p_dead_hosp", "p_dead_inf"),
  p_stat = c("mean", "low_95", "up_95"),
  p_sex = c("total", "male", "female")
)
}
\arguments{
\item{age_distr}{age distribution data-frame}

\item{p_type}{type of probablity to extract: p_hosp_inf = P(Hosp|Infection), etc}

\item{p_stat}{quantile of the original estimates used to compute probability}

\item{p_sex}{get estimates by sex or for total population}
}
\description{
Get age-sex specific probabilities estimated by Salje et al. 2020
}
\examples{
# overall propability of hospitalization given infection for France
age_dist_fr <- get_age_pop(iso = "FRA", format = "long")
get_p_Salje2020(age_dist_fr, p_type = "p_hosp_inf", p_stat = "mean", p_sex = "total")

}
\author{
Anton Camacho
}
