% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_severe_JHU.R
\name{get_p_severe_JHU}
\alias{get_p_severe_JHU}
\title{Get population-wide distribution of severe outcomes}
\usage{
get_p_severe_JHU(iso = "CHN", outcome = c("severe", "moderate", "mild"))
}
\arguments{
\item{iso}{iso_a3 of country of interest}

\item{outcome}{Outcome severity ("severe", "moderate", or "mild")}
}
\description{
This is set up to use population estimates from the World Populaiton
Prospects data, and age-specific probabilities of outcome severity from
Shenzhen
}
\examples{
# expected population distribution of moderate outcomes for Canada
get_p_severe_JHU(iso = "CAN", outcome = "moderate")

}
