% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_severe_JHU_model.R
\name{get_p_severe_JHU_model}
\alias{get_p_severe_JHU_model}
\title{Get the proprtion severe by age ditstr according to JHU assumptions}
\usage{
get_p_severe_JHU_model(age_distr, p)
}
\arguments{
\item{age_distr}{age distribution data-frame}

\item{p}{probability passed to quantile function}
}
\description{
Get the proprtion severe by age ditstr according to JHU assumptions
}
\examples{
# population age-distribution for France
pop_age_dist_france <- get_age_pop("FRA", format = "long")

# expected distribution of severe cases for France
get_p_severe_JHU_model(pop_age_dist_france, p = 0.5)

}
\author{
Flavio Finger
}
