% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_severe_neher.R
\name{get_p_severe_neher}
\alias{get_p_severe_neher}
\title{Get the age-adjusted proportion severe by country (giving ISO A3 code)
according to the assumptions taken by the Neher lab model}
\usage{
get_p_severe_neher(iso = "CHN")
}
\arguments{
\item{iso}{iso_a3 of country of interest (defaults to China, "CHN")}
}
\description{
Get the age-adjusted proportion severe by country (giving ISO A3 code)
according to the assumptions taken by the Neher lab model
}
\author{
Flavio Finger
}
