% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_severe_pop.R
\name{get_p_severe_pop}
\alias{get_p_severe_pop}
\title{Estimate proportion of severe cases adjusted for population structure}
\usage{
get_p_severe_pop(pr_age10)
}
\arguments{
\item{pr_age10}{proportion of population in 10 year age bins}
}
\description{
Estimate proportion of severe cases adjusted for population structure
}
\examples{
# population age-distribution for France
pop_age_dist_france <- get_age_pop("FRA")

# expected distribution of severe cases for France
get_p_severe_pop(pop_age_dist_france)

}
