% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_severe_pop_JHU.R
\name{get_p_severe_pop_JHU}
\alias{get_p_severe_pop_JHU}
\title{Estimate proportion of severe cases adjusted for population structure}
\usage{
get_p_severe_pop_JHU(age_distr)
}
\arguments{
\item{age_distr}{age distribution data-frame}
}
\description{
Estimate proportion of severe cases adjusted for population structure
}
\examples{
# population age-distribution for France
pop_age_dist_france <- get_age_pop("FRA", format = "long")

# expected distribution of severe cases for France
get_p_severe_pop_JHU(pop_age_dist_france)

}
