% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{get_pop_under12}
\alias{get_pop_under12}
\title{Population proportion under 12 for US states and counties}
\usage{
get_pop_under12(region)
}
\arguments{
\item{region}{A string with the state name, or the FIPS code}
}
\value{
State/county population as a numeric, or an error
}
\description{
Returns proportion of the 2019 census under 12 per state (proxy for county).
}
\examples{
get_pop_under12('Connecticut')
get_pop_under12('09009')

}
