% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_severe_age_Shenzhen.R
\name{get_severe_age_Shenzhen}
\alias{get_severe_age_Shenzhen}
\title{Model the posterior probability of Covid19 case severity by age group based on
data from Shenzen, China}
\usage{
get_severe_age_Shenzhen(outcome = list("severe", "moderate", "mild"))
}
\arguments{
\item{outcome}{Outcome category ("severe", "moderate", or "mild")}
}
\value{
data.frame
}
\description{
Estimates probability of being a severe case by age from Shenzen data.

Data in Bi et al. 2020
https://www.medrxiv.org/content/10.1101/2020.03.03.20028423v3.full.pdf

Adapted from https://github.com/HopkinsIDD/COVID19_refugees/
Now https://github.com/HopkinsIDD/covidSeverity ?
}
