% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_severity_neher.R
\name{get_severity_neher}
\alias{get_severity_neher}
\title{Get age severity data from Neher lab}
\usage{
get_severity_neher()
}
\description{
assumptions used by the Neher lab: https://neherlab.org/covid19/
accessed on 2020-04-02
"This table summarizes the assumptions on severity which are informed
by epidemiological and clinical observations in China. The first column
reflects our assumption on what fraction of infections are reflected
in the statistics from China, the following columns contain the
assumption on what fraction of the previous category deteriorates
to the next. These fields are editable and can be adjusted to different
assumptions. The last column is the implied infection fatality for
different age groups."
}
\author{
Flavio Finger
}
