% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{imperial}
\alias{imperial}
\title{Age-specific severity estimates from Imperial}
\format{
A data frame with 17 rows and 7 variables:
\describe{
  \item{age_group}{age group, in 5-year intervals}
  \item{p_hosp_inf}{probability of hospitalization given infection}
  \item{p_icu_hosp}{probability of ICU given hospitalization}
  \item{p_dead_icu}{probability of dying given ICU}
  \item{p_dead_nonicu}{probability of dying given non-ICU}
  \item{p_icu_inf}{probability of ICU given infection}
  \item{p_dead_inf}{probability of dying given infection}
}
}
\source{
\url{https://mrc-ide.github.io/global-lmic-reports/parameters.html}
}
\usage{
imperial
}
\description{
Age-specific severity estimates from Imperial
}
\keyword{datasets}
