% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input_cases}
\alias{input_cases}
\alias{input_deaths}
\alias{input_rr}
\alias{input_hosp}
\alias{input_boost}
\alias{input_vax}
\title{Input observational data}
\usage{
input_cases(data, type = "reported", lastCaseDate = NULL)

input_deaths(data, type = "reported", lastDeathDate = NULL)

input_rr(data, type = "reported")

input_hosp(data, type = "reported", lastHospDate = NULL, firstHospDate = NULL)

input_boost(data, type = "reported")

input_vax(data, type = "reported")
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}} as described below.}

\item{type}{A string, either \code{"reported"} or \code{"occurred"}.

\code{"reported"} applies to the following situations:
\itemize{
\item Case, death, hospitalizatons, booster or RR data where the count for a particular day
represents the number of cases or deaths that were reported publicly on
that day for the following week.
\item Cases, death, hospitalizatons, booster or RR data where the count for a particular day is the
number of cases or deaths received from hospital reports, or
diagnostic laboratories, accumulated across a particular region
}

\code{"occurred"} applies to the following situations:
\itemize{
\item Data where counts represent the number of cases or deaths that
actually occured on that day. For case data, this would be the day a
test was administered. For deaths, this would be the day an individual
died. Note that, if the deaths data represent the day an individual's
death was first reported as a SARS-Cov-2-related death, that data
should be passed with \code{type = "reported"}, instead.
}}
}
\description{
A family of functions used for inputting data into Covidestim.
}
\details{
There are two types of observational data that can be used with Covidestim:

\itemize{
\item Case data, detailing the number of new cases each week
\item Death data, detailing the number of confirmed Covid-19 deaths each
week
}

All input data to Covidestim is expected to be a
\code{\link[base]{data.frame}} of two variables. One variable, \code{date},
must be a vector of type \code{\link[base]{POSIXct}} or
\code{\link[base]{Date}}. The second column, \code{observations}, must be a
non-negative numeric vector.

Missing values in cases, hospitalizations, booster, RR and deaths data should be imputed before running the model.
The date range of all sets of data passed to \code{\link{covidestim}} must be
equivalent, with one observation each day, and no gaps in the data.
Assertions attempt to enforce this specification.
}
