% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconfig.R
\name{modelconfig}
\alias{modelconfig}
\title{Model configuration object. It can be modified through overloading
the addition operator}
\usage{
modelconfig(...)
}
\description{
Model configuration object. It can be modified through overloading
the addition operator
}
