% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconfig.R
\name{modelconfig_add.priors}
\alias{modelconfig_add.priors}
\title{Specialization for 'priors' classes.}
\usage{
\method{modelconfig_add}{priors}(rightside, leftside)
}
\description{
Splices in all the keys from a 'priors' obj into a 'modelconfig' (leftside)
object.
}
