% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{neher}
\alias{neher}
\title{Age-specific severity estimates from Neher lab}
\format{
A data frame with 9 rows and 9 variables:
\describe{
  \item{age}{age group, in 10-year intervals}
  \item{confirmed}{probability confirmed given infection}
  \item{severe}{probability severe case given confirmed}
  \item{critical}{probability of critical case given severe}
  \item{fatal}{probability of fatal case given critical}
  \item{p_hosp_inf}{probability of hospitalization given infection}
  \item{p_dead_hosp}{probability of dying given hospitalization}
  \item{p_dead_inf}{probability of dying given infection}
}
}
\source{
\url{https://covid19-scenarios.org/}
}
\usage{
neher
}
\description{
According to the source these estimaes are "informed by epidemiological and
clinical observations from China".
}
\keyword{datasets}
