% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconfig.R
\name{+.modelconfig}
\alias{+.modelconfig}
\title{An overloaded addition operator, dispatched on the type of the lhs argument.}
\usage{
\method{+}{modelconfig}(a, b)
}
\description{
Flips the order of the arguments and calls 'modelconfig_add' to enable type
matching on the rhs operand.
}
