% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pri_alpha}
\alias{pri_alpha}
\alias{pri_beta}
\title{Convert mean/variance stats to alpha/beta parameters of gamma distribution}
\usage{
pri_alpha(mv)

pri_beta(mv)
}
\arguments{
\item{mv}{A two-element numeric vector, where \code{[1]} represents mean
and \code{[2]} represents variance.}
}
\value{
A number representing the alpha or beta parameter
}
\description{
Using a method of moments approach, converts a sample with a specified
mean and variance to a parameterization of a gamma distribution
}
