% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_diagnosis}
\alias{priors_diagnosis}
\title{Priors on probability of diagnosis}
\usage{
priors_diagnosis(
  p_diag_if_sym = c(2, 2),
  p_diag_if_sev = c(1.5, 1),
  weekend_eff = c(2, 2)
)
}
\arguments{
\item{p_diag_if_sym}{A two-element numeric vector containing \code{c(alpha,
beta)} parameters of a Beta distribution modeling the probability of
diagnosed if symptomatic. The mean of this distribution should be less
the mean of the distribution parameterized by \code{p_diag_if_sev}.
Otherwise, a warning will be displayed.}

\item{p_diag_if_sev}{A two-element numeric vector containing \code{c(alpha,
beta)} parameters of a Beta distribution modeling the probability of
being diagnosed if severely ill.}

\item{weekend_eff}{A two-element numeric vector containing \code{c(alpha,
beta)} parameters of a Beta distribution modeling the effect of weekends on
probability of diagnosis. Weekend effect is assumed to impact probability of 
diagnosis at symptomatic and severe disease states equally.}
}
\value{
An S3 object of class 'priors'
}
\description{
This function returns a keyed list of priors related to probability of
diagnosis. Called with no arguments, the default values are returned.
}
\details{
Boundary avoiding priors are used by default; a weakly informative prior
is used for the probability of diagnosis if severely ill.
}
\examples{
cfg <- covidestim(ndays = 50) + priors_diagnosis(p_diag_if_sym = c(2, 2))
}
