% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_diagnosis_delays_scale}
\alias{priors_diagnosis_delays_scale}
\title{Priors on diagnostic delays}
\usage{
priors_diagnosis_delays_scale(dx_delay_sym = c(2, 2), dx_delay_sev = c(2, 2))
}
\arguments{
\item{dx_delay_sym}{A two element vector containing the \code{c(alpha,beta)} 
parameters of a Beta distribution modeling a scaleing factor. Delay to 
diagnosis for symptomatic cases is modeled as the fraction of time in 
the symptomatic disease state, scaled by this factor.}

\item{dx_delay_sev}{A two element vector containing the \code{c(alpha,beta)} 
parameters of a Beta distribution modeling a scaleing factor. Delay to 
diagnosis for severe cases is modeled as the fraction of time in 
the severe disease state, scaled by this factor.}
}
\value{
An S3 object of class 'priors'
}
\description{
This function returns a keyed list of priors related to delays in diagnosing
of cases and deaths.  Called with no arguments, the default values are 
returned.
}
\details{
Boundary avoiding priors are use by default.
}
\examples{
cfg <- covidestim(ndays = 50) + priors_diagnosis_delays_scale(dx_delay_sym = c(0.5, 0.1))
}
