% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_transitions}
\alias{priors_transitions}
\title{Priors for transitions}
\usage{
priors_transitions(
  p_sym_if_inf = c(5.143, 3.536),
  p_sym_if_inf_postO = c(2.5, 10),
  p_sev_if_sym = c(1.8854, 20.002),
  p_die_if_sev = c(28.239, 162.3),
  p_die_if_inf = c(15.915, 3167.1),
  p_die_if_inf_postO = c(2.55, 1594),
  ifr_decl_OR = c(12.031, 8.999)
)
}
\arguments{
\item{p_sym_if_inf}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters/hyperpriors of a Beta distribution modeling the
probability of becoming symptomatic if infectious.

Sources for default value:

\itemize{
\item \insertRef{byambasuren_estimating_2020}{covidestim}
\item \insertRef{mizumoto_estimating_2020}{covidestim}
\item \insertRef{nishiura_estimation_2020}{covidestim}
}}

\item{p_sev_if_sym}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters/hyperpriors of a Beta distribution modeling the
probability of transitioning into the "severe" health state if
symptomatic. "Severe" disease is defined as disease that would likely
require hospitalization.

Sources for default value:

\itemize{
\item \insertRef{cdc_covid-19_response_team_severe_2020}{covidestim}
\item \insertRef{verity_estimates_2020}{covidestim}
}}

\item{p_die_if_sev}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters/hyperpriors of a Beta distribution modeling the
probability of dying if in the "severe" health state.

Source for default value: \insertRef{cdc_covid-19_response_team_severe_2020}{covidestim}}

\item{p_die_if_inf}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters/hyperpriors of a Beta distribution modeling the
probability of dying if infected (e.g. the infection fatality rate). This
prior represents a national average value, which is later adjusted for
state and county-level factors, and to reflect higher fatality rates early
in the epidemic.

Source for default value: \insertRef{ODriscoll_Nature_2020}{covidestim}}

\item{ifr_decl_OR}{A two-element numeric vector containing \code{c(shape,
  rate)} parameters/hyperpriors of a Gamma distribution modeling the
elevated IFR in early 2020, relative to the present. Default value
represents an IFR 30\% higher in March of 2020, with 95\% interval 10-50\%
higher.}

\item{rr_decl_sev}{A two-element numeric vector containing \code{c(shape,
  rate)} parameters/hyperpriors of a Gamma distribution modeling the
decline in infection hospitalization ratio after Dec 2021 (omicron). Default value
represents an IFR in Dec 2021 7.5\% of before, with 95\% interval 4-10\%.}

\item{rr_decl_die}{A two-element numeric vector containing \code{c(shape,
  rate)} parameters/hyperpriors of a Gamma distribution modeling the
decline in IFR in after Dec 2021 (omicron). Default value
represents an IFR 2\% March of 2020, with 95\% interval 1-3\%.

Source for default value: Expert opinion}
}
\value{
An S3 object of class \code{priors}
}
\description{
This function returns a keyed list of priors related to progression to
various health states. Called with no arguments, the default values are
returned. Custom hyperpriors can be specified by passing values for the
parameters specified below. All parameters must be non-negative real
numbers. The return value of this function must be added to a
\code{covidestim} object using the addition operator (see examples).
}
\examples{
cfg <- covidestim(nweeks = 31, region = 'Connecticut', 
cum_p_inf_init = .5, start_p_imm = .5) + 
priors_transitions(p_sym_if_inf = c(0.5, 0.2))
}
