% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtEst.R
\name{RtEst.covidestim_result}
\alias{RtEst.covidestim_result}
\alias{RtNaiveEstim.covidestim_result}
\title{Calculate Effective Reproduction Number from Covidestim output}
\usage{
\method{RtEst}{covidestim_result}(ccr, window.length = 5, mean.si = 4.7, std.si = 2.9)

\method{RtNaiveEstim}{covidestim_result}(cc, window.length = 5, mean.si = 4.7, std.si = 2.9)
}
\arguments{
\item{ccr}{The result of calling \code{\link{run}}. An object of class
\code{covidestim_result}.}

\item{window.length}{The size of the moving window over which Rt should be 
estimated; defaults to 5 days. This number must be odd, so that the
interval's true center always falls on a complete day. It also must be less
than half the number of data days, i.e., \code{window.length < ndays/2}.}

\item{mean.si}{The mean serial interval to use in the Rt estimate; defaults
to 4.7 days.}

\item{std.si}{The standard deviation of the serial interval; defaults to 
2.9 days.}
}
\value{
A \code{data.frame} with columns \code{date}, \code{Rt},
  \code{Rt.lo}, \code{Rt.hi}, where \code{date} is a vector of
  \code{\link[base]{Date}} objects. Or, for \code{RtNaiveEstim}, columns
  \code{date}, \code{NaiveRt}, \code{NaiveRt.lo}, \code{NaiveRt.hi}
}
\description{
Creates a \code{data.frame} with Rt estimates over time and 95\% CIs.
}
\details{
This code makes use of the \code{EpiEstim} package, described here:

\insertRef{cori_new_2013}{covidestim}

The method used in \code{EpiEstim} is \code{"parametric_si"}, and a 5-day
sliding window is used by default. \strong{Warning messages specifying that 
you are \code{"estimating R too early"} are expected and should be ignored}.
}
