% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{run.covidestim}
\alias{run.covidestim}
\title{Run the Covidestim model}
\usage{
\method{run}{covidestim}(cc, cores = parallel::detectCores(), ...)
}
\arguments{
\item{cc}{A valid \code{\link{covidestim}} configuration}

\item{cores}{A number. How many cores to use to execute runs.}

\item{...}{Extra arguments to be passed to \code{\link[rstan]{sampling}}}
}
\value{
A S3 object of class \code{covidestim_result} containing the
  configuration used to run the model, the raw results, the extracted
  result as produced by \code{\link[rstan]{extract}}, and the summarized
  results as produced by \code{\link[rstan]{extract}}.
}
\description{
Calling \code{run()} with a \code{\link{covidestim}} object executes the
model and returns a result. \code{run} will attempt to run on as
many cores as appear to be available on the host machine, through calling
\code{\link[parallel]{detectCores}}. Model runtimes will range anywhere from
5-120 minutes.
}
\details{
When running in an interactive/TTY environment (like Rstudio, Radian, or the
R terminal), progress messages from \code{rstan} will be displayed,
indicatng how many iterations each chain has completed. When running in
other environments, for instance on a cluster, \code{rstan} will produce 
no output until the end of sampling.
}
