% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{runOptimizer.covidestim}
\alias{runOptimizer.covidestim}
\title{Run the Covidestim model using BFGS}
\usage{
\method{runOptimizer}{covidestim}(
  cc,
  cores = parallel::detectCores(),
  tries = 10,
  iter = 2000,
  timeout = 60,
  ...
)
}
\arguments{
\item{cc}{A valid \code{\link{covidestim}} configuration}

\item{cores}{A number. How many cores to use to execute runs. Multicore
execution is less important for \code{runOptimizer} than it is for
\code{\link{run}}.}

\item{tries}{The number of times to run the BFGS algorithm.}

\item{iter}{Passed to \code{\link[rstan]{optimizing}}.}

\item{timeout}{How long to let each run go for before killing it, in seconds.}

\item{...}{Arguments forwarded to \code{\link[rstan]{optimizing}}.}
}
\value{
An S3 object of class \code{covidestim_result}
}
\description{
In addition to NUTS sampling, you can fit the model using the BFGS algorithm.
This algorithm tries to maximize the mode of the posterior; it runs much
faster than NUTS, but won't return any confidence intervals. On
\url{https://covidestim.org}, BFGS is used to produce estimates for counties, and as
a fallback when NUTS fails to produce a timely and/or converged fit for state
data. Runtimes scale linearly to the value of \code{tries/cores}, but are
generally in the 30s-10min range. The same underlying model is used.
}
\details{
The BFGS algorithm is run \code{tries} times using \code{tries} different
seeds derived from \code{cc$seed}. Once all runs complete, the run with the
highest value of the log-posterior is selected and returned. BFGS
occasionally gets stuck; these runs are flagged and discarded before the
ranking begins.

A second method for fitting the model, using the NUTS algorithm, is available
as \link{run}. It provides CI's, but is significantly slower.\tabular{lllll}{
   \strong{Function} \tab \strong{Method} \tab \strong{CI's} \tab \strong{Speed} \tab \strong{Termination} \cr
   \code{run()} \tab NUTS \tab Yes \tab 30m-hours \tab Always, potentially with warnings, of which "treedepth" and "divergent transitions" are the most serious \cr
   \code{runOptimizer()} \tab BFGS \tab No \tab ~1-3min \tab Potentially with nonzero exit status (lack of convergence), or timeout (rare, gracefully handled internally) \cr
}
}
\examples{
cfg <- covidestim(nweeks = 31, region = 'Connecticut',
   pop = get_pop("Connecticut"),
   start_p_imm = get_imm_init("Connecticut")$start_p_imm,
   cum_p_inf_init = get_imm_init("Connecticut")$cum_p_inf_init) +
  input_cases(example_ct_data('cases')) +
  input_deaths(example_ct_data('deaths')) +
  input_rr(example_ct_data('RR')) + 
  input_hosp(example_ct_data('hosp')) +
  input_boost(example_ct_data('boost'))

\dontrun{
result <- runOptimizer(cfg, cores = 2)
}

}
\seealso{
\link{run.covidestim}
}
