% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{salje}
\alias{salje}
\title{Age-specific severity estimates from Salje et al. 2020}
\format{
A data frame with 96 rows and 5 variables:
\describe{
  \item{variable}{variable ("p_hosp_inf", "p_icu_hosp", "p_dead_hosp", or "p_dead_inf")}
  \item{age_group}{age group, in 10-year intervals except for first group 0-20}
  \item{sex}{sex ("male", "female", or "total")}
  \item{stat}{statistic, ("mean", "low_95", or "upp_95")}
  \item{probability}{probability associated with given variable}
  \item{quantile}{quantile corresponding to the relevant statistic}
}
}
\source{
Salje, H., Kiem, C.T., Lefrancq, N., Courtejoie, N., Bosetti, P.,
  Paireau, J., Andronico, A., Hoze, N., Richet, J., Dubost, C.L., and Le
  Strat, Y. (2020) Estimating the burden of SARS-CoV-2 in France. Science.
  \url{https://doi.org/10.1126/science.abc3517}
}
\usage{
salje
}
\description{
Estimates from Tables S1 and S2 of Salje et al. 2020.
}
\keyword{datasets}
