% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{shenzhen_samples}
\alias{shenzhen_samples}
\title{Samples of the posterior distribution of age-specific outcome probabilities
based on data from Schenzhen, China}
\format{A data frame with 96,000 rows and 4 variables:
\describe{
  \item{age_group}{age group, in 10-year intervals}
  \item{p}{sample from posterior probability distribution}
  \item{outcome}{outcome type}
  \item{model}{model used}
}}
\usage{
shenzhen_samples
}
\description{
Samples of the posterior distribution of age-specific outcome probabilities
based on data from Schenzhen, China
}
\keyword{datasets}
