% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.covidestim_result}
\alias{summary.covidestim_result}
\title{Summarize a Covidestim run}
\usage{
\method{summary}{covidestim_result}(ccr, include.before = TRUE, index = FALSE)
}
\arguments{
\item{ccr}{A \code{covidestim_result} object}

\item{include.before}{A logical scalar. Include estimations that fall in the
period before the first day of input data? (This period is of length
\code{nweeks_before} as passed to \code{covidestim}). If  \code{TRUE}, any
elements of variables which do not have values for this "before" period
will be represented as \code{NA}.}

\item{index}{A logical scalar. If \code{TRUE}, will include a variable
\code{index} in the output, with range \code{1:(nweeks_before + nweeks)}.}
}
\value{
A \code{data.frame} with the following variables:

  \itemize{
    \item \bold{\code{date}}

      Date as a \code{Date} vector.

    \item \bold{\code{deaths}}, \code{deaths} + (\code{_p2_5}, \code{_p25},
      \code{_p75}, \code{_p97_5})

      The number deaths estimated to occur on date 
      \code{date}. 

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{deaths_of_diagnosed}}, \code{deaths_of_diagnosed} +
      (\code{_p2_5}, \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of diagnosed deaths estimated to occur on date \code{date}.

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{diagnoses}}, \code{diagnoses} + (\code{_p2_5},
      \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of diagnoses estimated to occur on date \code{date}.

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{diagnoses_of_symptomatic}},
      \code{diagnoses_of_symptomatic} + (\code{_p2_5}, \code{_p25},
      \code{_p75}, \code{_p97_5})

      The number of diagnoses of symptomatic individuals estimated to occur
      on date \code{date}.

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{fitted_cases}}, \code{fitted_cases} + (\code{_p2_5},
      \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of modeled case reports for a date \code{date}. This takes
      the delay from diagnosis to report into account and thus is
      approximating how many case reports should exist for this date. This
      estimate is used to fit against the observed data (reported
      diagnoses).

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{fitted_deaths}}, \code{fitted_deaths} + (\code{_p2_5},
      \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of modeled death reports for a date \code{date}.  This
      takes the delay from death to report into account and thus is
      approximating how many death reports should exist for this date.  This
      estimate is used to fit against the observed data (reported deaths).
      
      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{fitted_hospitalizations}},
      \code{fitted_hospitalizations} + (\code{_p2_5}, \code{_p25},
      \code{_p75}, \code{_p97_5})

      The number of modeled hospitalization reports for a date \code{date}.
      This takes the delay from hospitalization to report into account and
      thus is approximating how many admission reports should exist for this
      date.  This estimate is used to fit against the observed data
      (reported hospital admissions).

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{infections_cumulative}}, \code{infections_cumulative}
      + (\code{_p2_5}, \code{_p25}, \code{_p75}, \code{_p97_5})

      The estimated cumulative number of infections on date \code{date}.
      This includes both first and repeat infections and can therefore
      exceed the population size.       

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{infections}}, \code{infections} + (\code{_p2_5},
      \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of infections estimated to occur on date \code{date}. 
      This includes both first and repeat infections.

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{infections_premiere}}, \code{infections_premiere} +
      (\code{_p2_5}, \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of first infections estimated to occur on date \code{date}.

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{r_t}}, \code{r_t} + (\code{_p2_5}, \code{_p25},
      \code{_p75}, \code{_p97_5})

      Estimate of the effective reproductive number (\eqn{R_t}).

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{seropositive_prvl}}, \code{seropositive_prvl} +
      (\code{_p2_5}, \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of individuals in the population who are modeled as being
      seropositive. This is an unreliable outcome that we don't recommend
      using. 
      
      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{severe}}, \code{severe} + (\code{_p2_5}, \code{_p25},
      \code{_p75}, \code{_p97_5})

      The number of transitions into the "severe" health state on date
      \code{date}. The "severe" state is defined as disease that would merit
      hospitalization. 
      
      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{susceptible_prvl}}, \code{susceptible_prvl} +
      (\code{_p2_5}, \code{_p25}, \code{_p75}, \code{_p97_5})

      The fraction of the population on date \code{date} that is susceptible
      to SARS-CoV-2 infection, i.e., the fraction of the population that has
      \emph{no} effective protection.

      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{symptomatic}}, \code{symptomatic} + (\code{_p2_5},
      \code{_p25}, \code{_p75}, \code{_p97_5})

      The number of modeled transitions of infected individuals into the
      infected, symptomatic health state on date \code{date}. This takes
      into account the probability of becoming symptomatic and the delay
      between infection and presentation of symptoms. 
      
      \emph{Median, 2.5\% interval, 25\% interval, 75\% interval, 97.5\%
        interval, ℝ}.

    \item \bold{\code{data_available}}

      Was there input data available for date \code{date}? This should be
      \code{TRUE}, except for the first month of estimates.

      \emph{logical}.
  }
}
\description{
Returns a \code{data.frame} summarizing a Covidestim model run. Note that if
\code{\link{runOptimizer.covidestim}} is used, all \code{*_p(2_5|25|75|97_5)}
variables will be \code{NA}-valued: the optimizer does not produce intervals.
}
