% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summaryEpi}
\alias{summaryEpi}
\title{Summarize internal parameters of a Covidestim run}
\usage{
summaryEpi(ccr)
}
\arguments{
\item{ccr}{A \code{covidestim_result} object}
}
\value{
A \code{data.frame} with the following variables:

\itemize{
\item \code{par}: Name of the Stan parameter

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\item \code{2.5\%}, \code{50\%}, \code{97.5\%}: Quantiles for the posterior
  distribution of \code{par}.
}\if{html}{\out{</div>}}

}
}
\description{
Returns a \code{data.frame} with a information on the posterior of a few
key Stan parameters.
}
