% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vanzandvoort}
\alias{vanzandvoort}
\title{Age-specific severity estimates from van Zandvoort et al. 2020}
\format{A data frame with 96 rows and 5 variables:
\describe{
  \item{age_group}{age group, in 10-year intervals}
  \item{p_hosp_clin}{probability of hospitalization given clinical}
  \item{p_icu_hosp}{probability of ICU given hospitalization}
  \item{p_dead_hosp}{probability of dying given hospitalization}
}}
\source{
van Zandvoort, K., Jarvis, C.I., Pearson, C., Davies, N.G., CMMID
  COVID-19 Working Group, Russell, T.W., Kucharski, A.J., Jit, M.J., Flasche,
  S., Eggo, R.M., and Checchi, F. (2020) Response strategies for COVID-19
  epidemics in African settings: a mathematical modelling study. medRxiv
  preprint. \url{https://doi.org/10.1101/2020.04.27.20081711}
}
\usage{
vanzandvoort
}
\description{
Estimates from Table S2 of van Zandvoort et al. 2020.
}
\keyword{datasets}
