% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{viz.covidestim_result}
\alias{viz.covidestim_result}
\title{Visualize results of a Covidestim run}
\usage{
\method{viz}{covidestim_result}(ccr, renderPDF = FALSE)
}
\arguments{
\item{ccr}{The result of calling \code{\link{run}}. An object of class
\code{covidestim_result}.}

\item{renderPDF}{If \code{FALSE}, figures are returned as a list of
\code{ggplot} objects. If a string, a \code{pdf} will be rendered to that
location, and and the list of \code{ggplot} objects will still be
returned.}
}
\value{
A list with the following keys:

\itemize{
\item \code{observedVsFitted} A \code{ggplot} of input cases/deaths data, and the
fit of the model to that data.
\item \code{infectionsAndCases} A \code{ggplot} of estimated daily infections,
compared to reported/observed cases.
}
}
\description{
Returns two graphs of Covidestim input and estimates
}
\seealso{
\code{\link{summary.covidestim_result}} for more details on the
quantities being plotted
}
