% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wpp_pop}
\alias{wpp_pop}
\title{Population estimates by country/region and 10-year age interval, derived from
the UN World Population Prospects 2019}
\format{
A data frame with 3765 rows and 16 variables:
\describe{
  \item{location}{country or region}
  \item{LocID}{location ID}
  \item{year}{year of populationestimates}
  \item{`0-9`}{population estimate for age group 0-9 years}
  \item{`10-19`}{population estimate for age group 10-19 years}
  \item{`20-29`}{population estimate for age group 20-29 years}
  \item{`30-39`}{population estimate for age group 30-39 years}
  \item{`40-49`}{population estimate for age group 40-49 years}
  \item{`50-59`}{population estimate for age group 50-59 years}
  \item{`60-69`}{population estimate for age group 60-69 years}
  \item{`70-79`}{population estimate for age group 70-79 years}
  \item{`80-89`}{population estimate for age group 80-89 years}
  \item{`90-99`}{population estimate for age group 90-99 years}
  \item{`100-109`}{population estimate for age group 100-109 years}
  \item{name}{location name}
  \item{iso_a3}{location ISO A3 code}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
wpp_pop
}
\description{
From \url{https://population.un.org/wpp/Download/Standard/Population/}

File: Population by Age Groups - Both Sexes (XLSX, 10.32 MB):
Quinquennial Population by Five-Year Age Groups - Both Sexes. De facto
population as of 1 July of the year indicated classified by five-year age
groups (0-4, 5-9, 10-14, ..., 95-99, 100+). (Accessed 2020-05-16)
}
\keyword{datasets}
