% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_IFR.R
\name{compare_IFR}
\alias{compare_IFR}
\title{Compile age-adjusted IFR estimates from different countries based on ensemble
estimates by O'Driscoll et al. 2020.}
\usage{
compare_IFR(countries = c("AFG", "PAK", "IRN"), p_sex = c("total",
  "male", "female"))
}
\arguments{
\item{countries}{A vector of ISO3 country codes used to extract age-specific population
estimates from the UN World Population Prospects 2019 dataset.}

\item{p_sex}{Use severity estimate for which sex (either "female", "male", or
"total")}
}
\value{
A data.frame with IFR estimates (mean and 95% CI) for different countries
}
\description{
Compile age-adjusted IFR estimates from different countries based on ensemble
estimates by O'Driscoll et al. 2020. The population age
distributions are either taken from the UN World Population
Prospects 2019 (WPP2019).
}
\examples{
compare_IFR(countries = c("AFG", "PAK", "IRN"))

}
\author{
Flavio Finger <flavio.finger@epicentre.msf.org>
}
