% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\docType{package}
\name{covidestim-package}
\alias{covidestim-package}
\title{The 'covidestim' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
