% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{covidestim}
\alias{covidestim}
\title{Configure a Covidestim run on a set of priors and input data}
\usage{
covidestim(
  nweeks,
  nweeks_before = 4,
  pop_size = 1e+12,
  OR = 1,
  p_hosp_nonsevere = 0.002,
  chains = 3,
  iter = 2000,
  nspl_rt_knotwidth = 1,
  thin = 1,
  seed = 42,
  adapt_delta = 0.98,
  max_treedepth = 14,
  window.length = 7,
  region
)
}
\arguments{
\item{nweeks}{A positive integer. The number of weeks of input data being
modeled. This should always be set to the number of weeks in your input
data.}

\item{nweeks_before}{A positive integer. How many weeks before the first week
of model data should be modeled? A higher number will produce estimates
that go farther back in time, however, those estimates will contain more
and more uncertainty.}

\item{pop_size}{A positive integer What is the population in the geography
being modelled? This sets the max susceptible population and becomes
important as the population ever infected approaches the population size.}

\item{chains}{The number of chains to use during NUTS sampling, as passed to
\code{\link[rstan]{sampling}}.}

\item{iter}{The number of iterations to run during NUTS sampling, as passed
to \code{\link[rstan]{sampling}}.}

\item{nspl_rt_knotwidth}{An integer. The knotwidth to use for the spline of
Rt.}

\item{thin}{A positive integer to specify period for saving samples, as
passed to \code{\link[rstan]{sampling}}. Modify this only if you intend
to inspect raw iterations data returned by \code{\link[rstan]{sampling}}.}

\item{seed}{A number. The random number generator seed for use in NUTS
sampling or in BFGS.}

\item{region}{A string. The FIPS code (for U.S. counties) or state name
(e.g. \code{New York}) being modeled. Required.}
}
\value{
An S3 object of type \code{covidestim}. This can be passed to
\code{\link{run.covidestim}} or \code{\link{runOptimizer.covidestim}} to execute the model, as
long as input data has been added (using the addition operator, see
example). This object can also be saved to disk using
\code{\link[base]{saveRDS}} to enable reproducibility across platforms or
sessions. However, note that Stan runs are only reproducible under very
specific conditions due to Stan's multi-language architecture. The
\code{print} method is overloaded to return to the user a summary of the
configuration, including prior values and the presence or absence of input
data.
}
\description{
\code{covidestim} returns a base configuration of the model with the default
set of priors, and no input data. This configuration, after adding input
data (see \code{\link{input_cases}}, \code{\link{input_deaths}},
\code{priors_*}), represents a valid model configuration that can be passed
to \code{\link{run.covidestim}} (for NUTS) or
\code{\link{runOptimizer.covidestim}} (for BFGS).
}
\examples{
cfg <- covidestim(nweeks = 31, region = 'Connecticut',
   pop = get_pop("Connecticut")) +
  input_cases(example_ct_data('cases')) +
  input_deaths(example_ct_data('deaths')) 
  

}
\seealso{
\url{https://mc-stan.org/docs/2_18/reference-manual/reproducibility-chapter.html}
}
