% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{covidestim}
\alias{covidestim}
\title{Configure a Covidestim run on a set of data and priors}
\usage{
covidestim(
  ndays,
  ndays_before = 28,
  pop_size = 1e+12,
  chains = 3,
  iter = 1500,
  thin = 1,
  seed = 42,
  adapt_delta = 0.92,
  max_treedepth = 12,
  window.length = 5
)
}
\arguments{
\item{ndays}{A positive integer. The number of days of input data being
modeled. This should always be set to the number of days in your input data.}

\item{ndays_before}{A positive integer. How many days before the first day
of model data should be modeled?}

\item{pop_size}{A positive real. What is the population in the geography 
being modelled? This sets the max susceptible population}

\item{chains}{The number of chains to use during sampling, as passed to
\code{\link[rstan]{sampling}}.}

\item{iter}{The number of iterations to run during sampling, as passed to
\code{\link[rstan]{sampling}}.}

\item{thin}{A positive integer to specify period for saving samples, as
passed to \code{\link[rstan]{sampling}}.}

\item{seed}{A number. The random number generator seed for use in sampling.}
}
\value{
An S3 object of type \code{covidestim}. This can be passed to 
  \code{\link{run}} to execute the model. This object can also be saved
  to disk using \code{\link[base]{saveRDS}} to enable reproducibility across
  platforms or sessions. The \code{print} method is overloaded to return
  to the user a summary of the configuration, including prior values and 
  the presence or absence of input data.
}
\description{
\code{covidestim} returns a base configuration of the model with the default
set of priors, and no input data. This configuration, after adding input
data (see \code{\link{input_cases}}, \code{\link{input_deaths}}, represents 
a valid model configuration that can be passed to \code{\link{run}}.
}
\examples{
covidestim(ndays = 50, seed = 42)
}
