% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{example_ct_data}
\alias{example_ct_data}
\title{Example CT case/death data}
\usage{
example_ct_data(type = "cases")
}
\arguments{
\item{type}{A string: \code{"cases"} or \code{"deaths"} or \code{"RR"} or
\code{"hosp"} or \code{"boost"}}
}
\value{
A tibble with two variables: \code{date}, a vector of \code{Date}
objects, and \code{observation}, a vector of doubles.
}
\description{
Returns a \code{\link[tibble]{tibble}} of Connecticut case, death, vaccine RR,
booster and hospitalizations data from December 2, 2021 to June 30, 2022.
Only one type of data will be returned at a time.
}
\examples{

cfg <- covidestim(nweeks = 32, region = 'Connecticut',
   pop = get_pop("Connecticut"),
   start_p_imm = get_imm_init("Connecticut")$start_p_imm,
   cum_p_inf_init = get_imm_init("Connecticut")$cum_p_inf_init) +
  input_cases(example_ct_data('cases')) +
  input_deaths(example_ct_data('deaths')) +
  input_rr(example_ct_data('RR')) + 
  input_hosp(example_ct_data('hospi')) +
  input_boost(example_ct_data('boost'))

\dontrun{
  result <- run(cfg)
}

}
