% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_pop.R
\name{get_age_pop}
\alias{get_age_pop}
\title{Get population distribution and aggregate it to 10 year age groups}
\usage{
get_age_pop(iso)
}
\arguments{
\item{iso}{country of interest}
}
\description{
This is set up to use population estimates from the World Populaiton
Prospects data
}
