% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifr.R
\name{get_ifr_raw}
\alias{get_ifr_raw}
\title{Estimated IFR for US counties and states}
\usage{
get_ifr_raw(region)
}
\arguments{
\item{region}{A string with the state name, or the FIPS code}
}
\value{
A data.frame with two variables: [date, value], where value is
  the odds ratio on VALUE HERE for that particular day. If \code{region}
  is specified incorrectly, an error will be thrown.
}
\description{
DETAILS HERE. Look in `data-raw/` for details on how this is generated.
}
\examples{
get_ifr_raw('Connecticut')
get_ifr_raw('09009')
}
