% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_Neher.R
\name{get_p_Neher}
\alias{get_p_Neher}
\title{Get age-adjusted probability of severe, ICU and death by country (giving ISO
A3 code) according to the assumptions taken by the Neher lab model}
\usage{
get_p_Neher(
  x,
  p_type = c("p_hosp_inf", "p_icu_hosp", "p_dead_hosp", "p_dead_inf")
)
}
\arguments{
\item{x}{Either an ISO3 country code used to extract age-specific population
estimates from the UN World Population Prospects 2019 dataset, \emph{or}, a
data.frame containing age categories in the first column and population
counts (or proportions) in the second column}

\item{p_type}{type of probablity to extract: p_hosp_inf = P(Hosp|Infection), etc}
}
\value{
Probability (scalar)
}
\description{
Get age-adjusted probability of severe, ICU and death by country (giving ISO
A3 code) according to the assumptions taken by the Neher lab model
}
\examples{
# mean Pr(hospitalization|infection) for Canada (ISO3 code "CAN"), taking age
# distribution from WPP2019
get_p_Neher(x = "CAN", p_type = "p_hosp_inf")

# use custom age-distribution
age_df <- data.frame(
  age = c("0-9", "10-19", "20-29", "30-39", "40-49", "50-59", "60-69", "70-79", "80+"),
  pop = c(1023, 1720, 2422, 3456, 3866, 4104, 4003, 3576, 1210),
  stringsAsFactors = FALSE
)
get_p_Neher(x = age_df, p_type = "p_hosp_inf")

}
\author{
Flavio Finger
}
