% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_Neher.R
\name{get_p_Neher}
\alias{get_p_Neher}
\title{Get age-adjusted probability of severe, ICU and death by country (giving ISO
A3 code) according to the assumptions taken by the Neher lab model}
\usage{
get_p_Neher(
  age_distr,
  p_type = c("p_hosp_inf", "p_icu_hosp", "p_dead_hosp", "p_dead_inf")
)
}
\arguments{
\item{age_distr}{age distribution data-frame}

\item{p_type}{type of probablity to extract: p_hosp_inf = P(Hosp|Infection), etc}
}
\description{
Get age-adjusted probability of severe, ICU and death by country (giving ISO
A3 code) according to the assumptions taken by the Neher lab model
}
\examples{
# overall propability of hospitalization given infection for France
age_dist_fr <- get_age_pop(iso = "FRA", format = "long")
get_p_Neher(age_dist_fr, p_type = "p_hosp_inf")
}
\author{
Flavio Finger
}
