% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_p_ODriscoll.R
\name{get_p_ODriscoll}
\alias{get_p_ODriscoll}
\title{Estimate Covid2019 outcome probabilities for a population given its age
distribution, and age-severity estimates from O'Driscoll al. (2020)}
\source{
Age-specific mortality and immunity patterns
of SARS-CoV-2 infection in 45 countries
Megan O'Driscoll, Gabriel Ribeiro Dos Santos,
Lin Wang, Derek A.T. Cummings, Andrew S Azman,
Juliette Paireau, Arnaud Fontanet, Simon Cauchemez, Henrik Salje
medRxiv 2020.08.24.20180851
\url{https://doi.org/10.1101/2020.08.24.20180851}
}
\usage{
get_p_ODriscoll(x, p_type = c("p_dead_inf"), p_stat = c("mean",
  "low_95", "up_95"), p_sex = c("total", "male", "female"))
}
\arguments{
\item{x}{Either an ISO3 country code used to extract age-specific population
estimates from the UN World Population Prospects 2019 dataset, \emph{or}, a
data.frame containing age categories in the first column and population
counts (or proportions) in the second column}

\item{p_type}{Outcome to estimate ("p_dead_inf" is the only option here)}

\item{p_stat}{Statistic of the severity estimates to use (either "mean",
"low_95", or "up_95")}

\item{p_sex}{Use severity estimate for which sex (either "female", "male", or
"total")}
}
\value{
Estimated outcome probability (scalar)
}
\description{
Estimate Covid19 outcome probabilities (death|infection), using
age-severity estimates from O'Driscoll al. (2020), and the population age
distribution for a given country, either taken from the UN World Population
Prospects 2019 (WPP2019) or directly supplied by the user.
}
\examples{
# mean Pr(death|infection) for Canada (ISO3 code "CAN"), taking age
# distribution from WPP2019
get_p_ODriscoll(x = "CAN", p_stat = "mean", p_sex = "total")

# use custom age-distribution
age_df <- data.frame(
  age = c("0-4", "5-9", "10-14", "15-19", "20-24", "25-29", "30-34", "35-39", "40-44", "45-49",
             "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80+"),
  pop = c(1023, 1720, 2422, 3456, 3866, 4104, 4003, 3576, 1210, 1023,
           1720, 2422, 3456, 3866, 4104, 4003, 3576),
  stringsAsFactors = FALSE
)

get_p_ODriscoll(x = age_df, p_type = "p_dead_inf", p_stat = "mean", p_sex = "total")

}
\author{
Anton Camacho

Patrick Barks <patrick.barks@epicentre.msf.org>

Flavio Finger <flavio.finger@epicentre.msf.org>
}
