% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{get_posterior_shenzhen}
\alias{get_posterior_shenzhen}
\title{Get posterior samples of prob(outcome | age) for Shenzhen}
\usage{
get_posterior_shenzhen(outcome = c("severe", "moderate", "mild"))
}
\arguments{
\item{outcome}{Outcome severity ("severe", "moderate", or "mild")}
}
\description{
Get posterior samples of prob(outcome | age) for Shenzhen
}
