% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{pri_gamma_shape}
\alias{pri_gamma_shape}
\alias{pri_gamma_rate}
\title{Convert mean/variance stats to shape/rate parameters of gamma distribution}
\usage{
pri_gamma_shape(mv)

pri_gamma_rate(mv)
}
\arguments{
\item{mv}{A two-element numeric vector, where \code{[1]} represents mean
and \code{[2]} represents variance.}
}
\value{
A number representing the shape or rate parameter
}
\description{
Using a method of moments approach, converts a sample with a specified
mean and variance to a parameterization of a gamma distribution
}
\details{
Implemented as:

\code{
  pri_gamma_shape <- function(mv) (mv[1]^2)/mv[2]
  pri_gamma_rate  <- function(mv) mv[1]/mv[2]
}
}
