% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_diagnosis}
\alias{priors_diagnosis}
\title{Priors on probability of diagnosis}
\usage{
priors_diagnosis(
  rr_diag_asy_vs_sym = c(2, 18),
  rr_diag_sym_vs_sev = c(2, 2),
  p_diag_if_sev = c(5, 2)
)
}
\arguments{
\item{rr_diag_asy_vs_sym}{A two-element numeric vector containing 
\code{c(alpha, beta)} parameters of a Beta distribution modeling the rate 
ratio of diagnosis among asymptomatic vs symptomatic, but not severe, 
infections.}

\item{p_diag_if_sev}{A two-element numeric vector containing \code{c(alpha,
beta)} parameters of a Beta distribution modeling the probability of
being diagnosed if severely ill.}

\item{rr_diag_sym_vs_asy}{A two-element numeric vector containing 
\code{c(alpha, beta)} parameters of a Beta distribution modeling the rate 
ratio of diagnosis at the symptomatic vs severe stage of infection.}
}
\value{
An S3 object of class 'priors'
}
\description{
This function returns a keyed list of priors related to probability of
diagnosis. Called with no arguments, the default values are returned.
}
\details{
Boundary avoiding priors are used by default; a weakly informative prior
is used for the probability of diagnosis if severely ill.
}
\examples{
cfg <- covidestim(ndays = 50) + priors_diagnosis(p_diag_if_sym = c(2, 2))
}
