% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_progression}
\alias{priors_progression}
\title{Fixed values on delay to progression}
\usage{
priors_progression(
  inf_prg_delay = c(3.413, 0.6051),
  sym_prg_delay = c(1.624, 0.2175),
  sev_prg_delay = c(2.061, 0.2277),
  asy_rec_delay = c(14, 2),
  pri_serial_i = c(129.1, 22.25)
)
}
\arguments{
\item{inf_prg_delay}{A two-element numeric vector containing \code{c(shape, rate)}
parameters of a Gamma distribution modeling the time from infection to 
sypmtom onset. 

Source for default value: \insertRef{lauer_incubation_2020}{covidestim}}

\item{sym_prg_delay}{A two-element numeric vector containing \code{c(shape, rate)}
parameters of a Gamma distribution modeling the time from symptom onset to 
severe disease. 

Source for default value: \insertRef{zhou_clinical_2020}{covidestim}}

\item{sev_prg_delay}{A two-element numeric vector containing
  \code{c(shape, rate)} parameters of a Gamma distribution modeling the
  time from severe symptoms to death. 
  
Source for default value: \insertRef{linton_incubation_2020}{covidestim}}

\item{asy_rec_delay}{A two-element numeric vector containing 
\code{c(shape,rate)} parameters of a Gamma distribution modeling the  
time from infection to recovery without symptom development.}

\item{pri_serial_i}{A two-element numeric vector containing 
\code{c(shape,rate)} parameters of a Gamma distribution modeling the 
serial interval, the average time between successive cases.}
}
\value{
An S3 object of class \code{priors}
}
\description{
This function returns a keyed list of values related to progression to
the infectious state. Called with no arguments, the default values are
returned. The following arguments can be passed to create different priors:
}
\examples{
cfg <- covidestim(ndays = 50) + priors_progression(inf_prg_delay = c(4, 1))
}
