% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_progression}
\alias{priors_progression}
\title{Fixed values on delay to progression}
\usage{
priors_progression(
  inf_prg_delay = c(3.413, 0.6051 * 7),
  sym_prg_delay = c(1.624, 0.2175 * 7),
  sev_prg_delay = c(2.061, 0.2277 * 7),
  asy_rec_delay = c(14, 2 * 7),
  pri_serial_i = c(34.615, 11.538 * 7),
  infect_dist = c(8, 1.241 * 7),
  seropos_dist = c(4.41, 0.042 * 7)
)
}
\arguments{
\item{inf_prg_delay}{A two-element numeric vector containing \code{c(shape, rate)}
parameters/hyperpriors of a Gamma distribution modeling the time from
infection to sypmtom onset.

Source for default value: \insertRef{lauer_incubation_2020}{covidestim}}

\item{sym_prg_delay}{A two-element numeric vector containing \code{c(shape, rate)}
parameters/hyperpriors of a Gamma distribution modeling the time (in
number of days) from symptom onset to severe disease.

Source for default value: \insertRef{zhou_clinical_2020}{covidestim}}

\item{sev_prg_delay}{A two-element numeric vector containing
\code{c(shape, rate)} parameters/hyperpriors of a Gamma distribution
modeling the time (in number of days) from onset of severe symptoms to
death.

Source for default value: \insertRef{linton_incubation_2020}{covidestim}}

\item{asy_rec_delay}{A two-element numeric vector containing
\code{c(shape,rate)} parameters/hyperpriors of a Gamma distribution
modeling the time (in number of days) from infection to recovery for
individuals who never develop symptoms.}

\item{pri_serial_i}{A two-element numeric vector containing
\code{c(shape,rate)} parameters/hyperpriors of a Gamma distribution
modeling the serial interval, the average time (in number of days)
between successive cases.}

\item{infect_dist}{A two-element numeric vector containing
\code{c(shape,rate)} parameters/hyperpriors of a Gamma distribution modeling the
changes of infectiousness, following initial infection. The PDF of this
distribution can be thought of as describing how infectious an individual
is, as a function of time (in days).}

\item{seropos_dist}{A two-element numeric vector containing
\code{c(shape,rate)} parameters/hyperpriors of a Gamma distribution modeling the
time (in days) to seroreversion, following initial infection. We do not
recommend modifying this value, as it is an experimental feature, and not
used in the core model.}
}
\value{
An S3 object of class \code{priors}
}
\description{
This function returns a keyed list of priors related to progression to
the infectious state. Called with no arguments, the default values are
returned. Custom hyperpriors can be specified by passing values for the
parameters specified below. The return value of this function must be added
to a \code{covidestim} object using the addition operator (see examples). The
following arguments can be passed to create different priors:
}
\examples{
cfg <- covidestim(ndays = 50, region = 'New York') + priors_progression(inf_prg_delay = c(4, 1))
}
