% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_progression}
\alias{priors_progression}
\title{Fixed values on delay to progression}
\usage{
priors_progression(
  inf_prg_delay = c(5.202, 0.946),
  sym_prg_delay = c(5.147, 0.468),
  sev_prg_delay = c(2.383, 0.27)
)
}
\arguments{
\item{inf_prg_delay}{A two-element numeric vector containing \code{c(shape, scale)}
parameters of a Gamma distribution modeling the time from infection to 
sypmtom onset. 

Source for default value: \insertRef{lauer_incubation_2020}{covidestim}}

\item{sym_prg_delay}{A two-element numeric vector containing \code{c(shape, scale)}
parameters of a Gamma distribution modeling the time from symptom onset to 
severe disease. 

Source for default value: \insertRef{zhou_clinical_2020}{covidestim}}

\item{sev_prg_delay}{A two-element numeric vector containing
  \code{c(shape, scale)} parameters of a Gamma distribution modeling the
  time from severe symptoms to death. 

Source for default value: \insertRef{linton_incubation_2020}{covidestim}}
}
\value{
An S3 object of class \code{priors}
}
\description{
This function returns a keyed list of values related to progression to
the infectious state. Called with no arguments, the default values are
returned. The following arguments can be passed to create different priors:
}
\examples{
cfg <- covidestim(ndays = 50) + priors_progression(inf_prg_delay = c(4, 1))
}
