% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_transitions}
\alias{priors_transitions}
\title{Priors for transitions}
\usage{
priors_transitions(
  p_sym_if_inf = c(44.9, 9.9),
  p_sev_if_sym = c(11.9, 47.7),
  p_die_if_sev = c(3, 97),
  p_die_if_inf = c(72.147, 10976)
)
}
\arguments{
\item{p_sym_if_inf}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of being
  symptomatic if infectious. 

  Sources for default value: 
  
  \itemize{
    \item \insertRef{byambasuren_estimating_2020}{covidestim}
    \item \insertRef{mizumoto_estimating_2020}{covidestim}
    \item \insertRef{nishiura_estimation_2020}{covidestim}
  }}

\item{p_sev_if_sym}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of
  severe if symptomatic. 

  Sources for default value: 

  \itemize{
    \item \insertRef{cdc_covid-19_response_team_severe_2020}{covidestim}
    \item \insertRef{verity_estimates_2020}{covidestim}
  }}

\item{p_die_if_sev}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of dying
  if severely ill. 

  Source for default value: \insertRef{cdc_covid-19_response_team_severe_2020}{covidestim}}

\item{p_die_if_inf}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of dying
  if infected (e.g. the infection fatality rate). 

  Source for default value: \insertRef{basu_estimating_2020}{covidestim}}
}
\value{
An S3 object of class \code{priors}
}
\description{
This function returns a keyed list of priors related to progression to
various health states. Called with no arguments, the default values are
returned. All parameters must be non-negative real numbers.
}
\examples{
cfg <- covidestim(ndays = 50) + priors_transitions(p_sym_if_inf = c(0.5, 0.2))
}
