% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors_transitions}
\alias{priors_transitions}
\title{Priors for transitions}
\usage{
priors_transitions(
  p_sym_if_inf = c(5.143, 3.536),
  p_sev_if_sym = c(1.8854, 20.002),
  p_die_if_sev = c(28.239, 162.3),
  p_die_if_inf = c(15.915, 3167.1),
  ifr_decl_OR = c(9.1357, 29.339)
)
}
\arguments{
\item{p_sym_if_inf}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of being
  symptomatic if infectious. 

  Sources for default value: 
  
  \itemize{
    \item \insertRef{byambasuren_estimating_2020}{covidestim}
    \item \insertRef{mizumoto_estimating_2020}{covidestim}
    \item \insertRef{nishiura_estimation_2020}{covidestim}
  }}

\item{p_sev_if_sym}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of
  severe if symptomatic. 

  Sources for default value: 

  \itemize{
    \item \insertRef{cdc_covid-19_response_team_severe_2020}{covidestim}
    \item \insertRef{verity_estimates_2020}{covidestim}
  }}

\item{p_die_if_sev}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of dying
  if severely ill. 

  Source for default value: \insertRef{cdc_covid-19_response_team_severe_2020}{covidestim}}

\item{p_die_if_inf}{A two-element numeric vector containing \code{c(alpha,
  beta)} parameters of a Beta distribution modeling the probability of dying
  if infected (e.g. the infection fatality rate). This prior represents a 
  national average value, which is adjusted for state and county-level 
  factors, and to reflect higher fatality rates early in the epidemic.

  Source for default value: \insertRef{ODriscoll_Nature_2020}{covidestim}}

\item{ifr_decl_OR}{A two-element numeric vector containing \code{c(shape,
  rate)} parameters of a Gamma distribution modeling the elevated IFR in early
  2020, relative to the present. Default value represents an IFR 30% higher
  in March of 2020, with 95% interval 10-50% higher.

  Source for default value: Expert opinion}
}
\value{
An S3 object of class \code{priors}
}
\description{
This function returns a keyed list of priors related to progression to
various health states. Called with no arguments, the default values are
returned. All parameters must be non-negative real numbers.
}
\examples{
cfg <- covidestim(ndays = 50) + priors_transitions(p_sym_if_inf = c(0.5, 0.2))
}
