% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidestim.R
\name{run.covidestim}
\alias{run.covidestim}
\title{Run the Covidestim model using NUTS}
\usage{
\method{run}{covidestim}(cc, cores = parallel::detectCores(), ...)
}
\arguments{
\item{cc}{A valid \code{\link{covidestim}} configuration}

\item{cores}{A number. How many cores to use to execute runs.}

\item{...}{Extra arguments, passed on to \code{\link[rstan]{sampling}}}
}
\value{
A S3 object of class \code{covidestim_result} containing the
configuration used to run the model, the raw Stan results, the extracted
result as produced by \code{\link[rstan]{extract}}, and the summarized
results as produced by \code{\link[rstan]{extract}}.
}
\description{
Calling \code{run()} with a \code{\link{covidestim}} object executes the
model and returns a result. \code{run} will attempt to run on as
many cores as appear to be available on the host machine, through calling
\code{\link[parallel]{detectCores}}. Model runtimes will range anywhere from
30 minutes to 12 hours, with the cumulative number of cases appearing to be
a strong correlate to longer runtime.
}
\details{
When running in an interactive/TTY environment (like Rstudio, Radian, or the
R terminal), progress messages from \code{rstan} will be displayed,
indicatng how many iterations each chain has completed. When running in
other environments, for instance on a cluster, \code{rstan} will produce
no output until the end of sampling.

The sampler may return warnings upon completion. In general, "treedepth" and
"divergent transitions" are the most serious. Use caution when interpreting
results which were accompanied by these messages. See \href{https://mc-stan.org/misc/warnings.html}{mc-stan.org} for
a detailed guide to Stan's most common warning messages.

A second method for fitting the model, using the BFGS algorithm, is available
as \link{runOptimizer}. It is significantly faster, but lacks CI's.\tabular{lllll}{
   \strong{Function} \tab \strong{Method} \tab \strong{CI's} \tab \strong{Speed} \tab \strong{Termination} \cr
   \code{run()} \tab NUTS \tab Yes \tab 30m-hours \tab Always, potentially with warnings, of which "treedepth" and "divergent transitions" are the most serious \cr
   \code{runOptimizer()} \tab BFGS \tab No, \code{*.(lo|hi) == NA} \tab ~1-3min \tab Potentially with nonzero exit status (rare), or timeout (rare, gracefully handled internally) \cr
}
}
\examples{
# Note that this configuration is improper as it uses New York City
# case/death data, but uses Manhattan's FIPS code ('36061') and population size.
# (for demonstration purposes only!)
cfg <- covidestim(ndays = 120, seed = 42, region = '36061', pop_size = 1.63e6) +
  input_cases(example_nyc_data('cases')) +
  input_deaths(example_nyc_data('deaths'))

\dontrun{
result <- run(cfg, cores = 2)
}

}
\seealso{
\code{\link{runOptimizer}}
}
