% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{salje}
\alias{salje}
\title{Age-specific severity estimates from Salje et al. 2020}
\format{A data frame with 96 rows and 5 variables:
\describe{
  \item{age_group}{age group, in 10-year intervals except for first group 0-19}
  \item{sex}{sex ("male", "female", or "total")}
  \item{stat}{statistic ("mean", "low_95", or "upp_95")}
  \item{quantile}{quantile corresponding to the relevant statistic}
  \item{p_hosp_inf}{probability of hospitalization given infection}
  \item{p_icu_hosp}{probability of ICU given hospitalization}
  \item{p_dead_hosp}{probability death given hospitalization}
  \item{p_dead_inf}{probability of death given infection}
}}
\source{
Salje, H., Kiem, C.T., Lefrancq, N., Courtejoie, N., Bosetti, P.,
  Paireau, J., Andronico, A., Hoze, N., Richet, J., Dubost, C.L., and Le
  Strat, Y. (2020) Estimating the burden of SARS-CoV-2 in France. Science.
  \url{https://doi.org/10.1126/science.abc3517}
}
\usage{
salje
}
\description{
Estimates from Tables S1 and S2 of Salje et al. 2020.
}
\keyword{datasets}
