% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.covidestim_result}
\alias{summary.covidestim_result}
\title{Summarize a Covidestim run}
\usage{
\method{summary}{covidestim_result}(ccr, include.before = TRUE, index = FALSE)
}
\arguments{
\item{ccr}{A \code{covidestim_result} object}

\item{include.before}{A logical scalar. Include estimations that fall in the
period before the first day of input data? (This period is of length
\code{ndays_before} as passed to \code{covidestim}). If  \code{TRUE}, any
elements of variables which do not have values for this "before" period
will be represented as \code{NA}.}

\item{index}{A logical scalar. If \code{TRUE}, will include a variable
\code{index} in the output, with range \code{1:(ndays_before + ndays)}.}
}
\value{

}
\description{
Returns a \code{data.frame} summarizing a Covidestim model run. Note that if
\code{\link{runOptimizer.covidestim}} is used, all \code{*.(lo|hi)}
variables will be \code{NA}-valued, because BFGS does not generate
confidence intervals.
}
